requirements.txt
requests
pillow
pytesseract
regex